+++ Job3/include/kernelRedux.h
@@ -1,4 +1,4 @@
 #ifndef __KERNEL_REDUX_H__
 #define __KERNEL_REDUX_H__
-__global__ void row_kernel(float*,float*);
+__global__ void row_kernel(double*,double*);
 #endif
